fix: validate sendconfirmation before password if possible
