+++ packages/shared/lib/core/error/classes/send/invalid-expiration-date-time-error.class.ts
@@ -5,7 +5,7 @@ export class InvalidExpirationDateTimeError extends BaseError {
     constructor() {
         super({
             message: localize('error.send.invalidExpirationDateTime'),
-            logError: true,
+            logError: false,
         })
     }
 }
