+++ packages/shared/lib/core/wallet/actions/sendOutput.ts
@@ -1,6 +1,4 @@
 import { selectedAccount } from '@core/account'
-import { InsufficientFundsForStorageDepositError, InvalidExpirationDateTimeError } from '@core/error'
-import { convertUnixTimestampToDate, isValidExpirationDateTime } from '@core/utils'
 import { OutputTypes } from '@iota/types'
 import { OutputOptions, TransactionOptions } from '@iota/wallet'
 import { isTransferring } from '@lib/wallet'
@@ -11,7 +9,6 @@ import { addActivityToAccountActivitiesInAllAccountActivities } from '../stores'
 export async function sendOutput(outputOptions: OutputOptions, output: OutputTypes): Promise<void> {
     try {
         isTransferring.set(true)
-        validate(outputOptions, output)
         const account = get(selectedAccount)
         const transferOptions: TransactionOptions = {
             remainderValueStrategy: { strategy: 'ReuseAddress', value: null },
@@ -28,20 +25,5 @@ export async function sendOutput(outputOptions: OutputOptions, output: OutputTyp
     } catch (err) {
         isTransferring.set(false)
         throw err
-    }
-}
-
-function validate(outputOptions: OutputOptions, outputTypes: OutputTypes): void {
-    const parseNumber = (value: string) => parseInt(value, 10) ?? 0
-    const amount = parseNumber(outputTypes?.amount)
-    const balance = parseNumber(get(selectedAccount)?.balances?.available)
-    // const storageDeposit = calculateStorageDepositFromOutput(outputTypes, amount)
-    const storageDeposit = 10000000000
-    const expirationDateTime = convertUnixTimestampToDate(outputOptions?.unlocks?.expirationUnixTime)
-
-    if (balance < amount + storageDeposit) {
-        throw new InsufficientFundsForStorageDepositError(storageDeposit + amount - balance, '')
-    } else if (expirationDateTime && !isValidExpirationDateTime(expirationDateTime)) {
-        throw new InvalidExpirationDateTimeError()
     }
 }
