+++ packages/shared/lib/core/wallet/utils/index.ts
@@ -11,4 +11,5 @@ export * from './isSubjectInternal'
 export * from './parseRawAmount'
 
 export * from './outputs'
+export * from './send'
 export * from './transactions'
