+++ packages/shared/lib/core/wallet/utils/send/index.ts
@@ -1,0 +1,1 @@
+export * from './validateSendConfirmation'
